+++ DB/DBC/standard/business/src/main/java/jp/co/ndensan/reams/db/dbc/business/core/basic/KagoKetteiMeisaiBuilder.java
@@ -73,7 +73,7 @@ public class KagoKetteiMeisaiBuilder {
      * @param 履歴番号 履歴番号
      * @return {@link KagoKetteiMeisaiBuilder}
      */
-    public KagoKetteiMeisaiBuilder set履歴番号(Decimal 履歴番号) {
+    public KagoKetteiMeisaiBuilder set履歴番号(int 履歴番号) {
         requireNonNull(履歴番号, UrSystemErrorMessages.値がnull.getReplacedMessage("履歴番号"));
         entity.setRirekiNo(履歴番号);
         return this;
